+++ requirements.txt
@@ -25,7 +25,7 @@ mock==5.0.1 ; python_version >= "3.9" and python_version < "4.0"
 mutagen==1.46.0 ; python_version >= "3.9" and python_version < "4.0"
 pathvalidate==2.5.2 ; python_version >= "3.9" and python_version < "4.0"
 plumbum==1.8.1 ; python_version >= "3.9" and python_version < "4.0"
-prometheus-client==0.15.0 ; python_version >= "3.9" and python_version < "4.0"
+prometheus-client==0.16.0 ; python_version >= "3.9" and python_version < "4.0"
 pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation != "CPython"
 pycryptodomex==3.17 ; python_version >= "3.9" and python_version < "4.0"
 pysocks==1.7.1 ; python_version >= "3.9" and python_version < "4"
@@ -52,4 +52,4 @@ unicode-slugify==0.1.5 ; python_version >= "3.9" and python_version < "4.0"
 unidecode==1.3.6 ; python_version >= "3.9" and python_version < "4.0"
 urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4"
 websockets==10.4 ; python_version >= "3.9" and python_version < "4.0"
-yt-dlp==2023.1.6 ; python_version >= "3.9" and python_version < "4.0"
+yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git@master ; python_version >= "3.9" and python_version < "4.0"
