--- scdlbot/exceptions.py
@@ -1,122 +1,0 @@
-class Error(Exception):
-    """Base class for exceptions in this module."""
-
-
-class FileNotSupportedError(Error):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self, file_format):
-        self.file_format = file_format
-
-
-class FileTooLargeError(Error):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self, file_size):
-        self.file_size = file_size
-
-
-class FileSplittedPartiallyError(Error):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self, file_parts):
-        self.file_parts = file_parts
-
-
-class FileNotConvertedError(Error):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self):
-        pass
-
-
-class FileSentPartiallyError(Error):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self, sent_audio_ids):
-        self.sent_audio_ids = sent_audio_ids
-
-
-class URLError(Error):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self):
-        self.status = ""
-
-
-class URLDirectError(URLError):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self):
-        self.status = "direct"
-
-
-class URLCountryError(URLError):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self):
-        self.status = "country"
-
-
-class URLLiveError(URLError):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self):
-        self.status = "live"
-
-
-class URLTimeoutError(URLError):
-    """Exception raised for errors in the input.
-
-    Attributes:
-        expression -- input expression in which the error occurred
-        message -- explanation of the error
-    """
-
-    def __init__(self):
-        self.status = "timeout"
